extract expr2smvt into a separate header file
